Checking in wip - discovery refactor
